b2note_api/requirements.txt
asn1crypto==0.24.0
Authlib==0.11
cachetools==3.1.1
Cerberus==1.3.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Eve==0.9.1
Eve-Swagger==0.0.11
Events==0.3
Flask==1.0.3
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pymongo==3.8.0
redis==3.2.1
requests==2.22.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
